added type caste to avoid warning and fixed line intend

+++ src/getdelim.c
@@ -64,72 +64,72 @@ ssize_t getline(char **restrict lineptr, size_t *restrict n,
 ssize_t getdelim(char **restrict lineptr, size_t *restrict n, int delimiter,
                  FILE *restrict stream)
 {
-	char *buf, *pos;
-	int c;
-	ssize_t bytes;
+  char *buf, *pos;
+  int c;
+  ssize_t bytes;
 
-	if (lineptr == NULL || n == NULL) {
-		errno = EINVAL;
-		return -1;
-	}
-	if (stream == NULL) {
-		errno = EBADF;
-		return -1;
-	}
+  if (lineptr == NULL || n == NULL) {
+    errno = EINVAL;
+    return -1;
+  }
+  if (stream == NULL) {
+    errno = EBADF;
+    return -1;
+  }
 
-	/* resize (or allocate) the line buffer if necessary */
-	buf = *lineptr;
-	if (buf == NULL || *n < _GETDELIM_MINLEN) {
-		buf = realloc(*lineptr, _GETDELIM_GROWBY);
-		if (buf == NULL) {
-			/* ENOMEM */
-			return -1;
-		}
-		*n = _GETDELIM_GROWBY;
-		*lineptr = buf;
-	}
+  /* resize (or allocate) the line buffer if necessary */
+  buf = *lineptr;
+  if (buf == NULL || *n < _GETDELIM_MINLEN) {
+    buf = realloc(*lineptr, _GETDELIM_GROWBY);
+    if (buf == NULL) {
+      /* ENOMEM */
+      return -1;
+    }
+    *n = _GETDELIM_GROWBY;
+    *lineptr = buf;
+  }
 
-	/* read characters until delimiter is found, end of file is reached, or an
-	   error occurs. */
-	bytes = 0;
-	pos = buf;
-	while ((c = getc(stream)) != EOF) {
-		if (bytes + 1 >= SSIZE_MAX) {
-			errno = EOVERFLOW;
-			return -1;
-		}
-		bytes++;
-		if (bytes >= *n - 1) {
-			buf = realloc(*lineptr, *n + _GETDELIM_GROWBY);
-			if (buf == NULL) {
-				/* ENOMEM */
-				return -1;
-			}
-			*n += _GETDELIM_GROWBY;
-			pos = buf + bytes - 1;
-			*lineptr = buf;
-		}
+  /* read characters until delimiter is found, end of file is reached, or an
+       error occurs. */
+  bytes = 0;
+  pos = buf;
+  while ((c = getc(stream)) != EOF) {
+    if (bytes + 1 >= SSIZE_MAX) {
+      errno = EOVERFLOW;
+      return -1;
+    }
+    bytes++;
+    if ((size_t)bytes >= *n - 1) {
+      buf = realloc(*lineptr, *n + _GETDELIM_GROWBY);
+      if (buf == NULL) {
+        /* ENOMEM */
+        return -1;
+      }
+      *n += _GETDELIM_GROWBY;
+      pos = buf + bytes - 1;
+      *lineptr = buf;
+    }
 
-		*pos++ = (char) c;
-		if (c == delimiter) {
-			break;
-		}
-	}
+    *pos++ = (char) c;
+    if (c == delimiter) {
+      break;
+    }
+  }
 
-	if (ferror(stream) || (feof(stream) && (bytes == 0))) {
-		/* EOF, or an error from getc(). */
-		return -1;
-	}
+  if (ferror(stream) || (feof(stream) && (bytes == 0))) {
+    /* EOF, or an error from getc(). */
+    return -1;
+  }
 
-	*pos = '\0';
-	return bytes;
+  *pos = '\0';
+  return bytes;
 }
 
 
 ssize_t getline(char **restrict lineptr, size_t *restrict n,
                 FILE *restrict stream)
 {
-	return getdelim(lineptr, n, '\n', stream);
+  return getdelim(lineptr, n, '\n', stream);
 }
 
 
@@ -138,12 +138,12 @@ ssize_t getline(char **restrict lineptr, size_t *restrict n,
 /* TODO: this isn't a very extensive test. */
 int main(void)
 {
-	char *line = NULL;
-	size_t n = 0;
-	while (getline(&line, &n, stdin) > 0) {
-		printf("%s", line);
-	}
-	return 0;
+  char *line = NULL;
+  size_t n = 0;
+  while (getline(&line, &n, stdin) > 0) {
+    printf("%s", line);
+  }
+  return 0;
 }
 
 #endif
